Use 'bindings' modules to load posix.node.
Depending on how gyp and node are configured, they can build and try to
load native modules in a number of different locations. The current
hardcoded 'Release' dir is not always correct resulting in exceptions
while loading. In particular, it doesn't handle the case where node is
built in 'Debug' mode. The 'bindings' module handles the common cases of
this issue.