version 2.0.0
IMPORTANT NOTE: version 2.0.0 and greater introduce a new dependency
to "nan", which is a library providing node.js C++ API compatibility
support for both node.js versions prior and after 1.2.0, which is
the point where a major compatibility break in the interface was
introduced.

Also io.js supported starting from this version and onwards.

node.js 0.6.x support is dropped.